refactor: move common source bus access methods to source-client.js
