+++ src/source/get.js
@@ -11,8 +11,9 @@
  */
 import { createErrorResponse } from '../contentbus/utils.js';
 import { listFolder } from './folder.js';
-import { accessSourceFile, getS3KeyFromInfo } from './utils.js';
-import { getOrListVersions, VERSION_FOLDER } from './versions.js';
+import { accessSourceFile, VERSION_FOLDER } from './source-client.js';
+import { getS3KeyFromInfo } from './utils.js';
+import { getOrListVersions } from './versions.js';
 
 async function accessSource(context, info, headRequest) {
   if (info.rawPath.endsWith('/')) {
