+++ src/source/put.js
@@ -10,194 +10,16 @@
  * governing permissions and limitations under the License.
  */
 import { Response } from '@adobe/fetch';
-import processQueue from '@adobe/helix-shared-process-queue';
-import { HelixStorage } from '@adobe/helix-shared-storage';
-import { ulid } from 'ulid';
 import { createErrorResponse } from '../contentbus/utils.js';
-import { StatusCodeError } from '../support/StatusCodeError.js';
 import { checkConditionals } from './header-utils.js';
+import { copyFolder, copyDocument, storeSourceFile } from './source-client.js';
 import {
   contentTypeFromExtension,
   getDocPathFromS3Key,
   getS3KeyFromInfo,
   getS3Key,
-  getDocID,
   getValidPayload,
-  storeSourceFile,
-  MAX_SOURCE_BUCKET_RETRY,
 } from './utils.js';
-import { postVersion } from './versions.js';
-
-/**
- * Copy an S3 object and handle conflichts.
- *
- * @param {import('../support/AdminContext').AdminContext} context context
- * @param {string} srcKey source S3 key
- * @param {string} destKey destination S3 key
- * @param {boolean} move true if this is a move operation
- * @param {object} initialOpts metadata options for the copy operation
- * @param {object} collOpts collision options (e.g { collision: 'overwrite' } ),
- * these collision options are used to handle conflicts when copying the source to the
- * destination.
- * - 'overwrite' - overwrite the destination if it exists, but create a version of
- * the destination first.
- * - 'unique' - append a (base-36) encoded timestamp to the destination key to make it unique.
- * These timestamps are alphabetically sortable.
- */
-async function copyWithRetry(
-  context,
-  srcKey,
-  destKey,
-  move,
-  initialOpts,
-  collOpts,
-) {
-  const bucket = HelixStorage.fromContext(context).sourceBus();
-  let destinationKey = destKey;
-  let opts = initialOpts;
-
-  // We start with assuming that there is nothing at the destination, the happy path
-  let copyOpts = { IfNoneMatch: '*' };
-
-  const maxRetry = context.attributes.maxSourceBucketRetry ?? MAX_SOURCE_BUCKET_RETRY;
-  let attempt = 0;
-  while (true) {
-    try {
-      const allOpts = { copyOpts, ...opts };
-      // eslint-disable-next-line no-await-in-loop
-      await bucket.copy(srcKey, destinationKey, allOpts);
-
-      break; // copy was successful, break out of the loop - we're done!
-    } catch (e) {
-      attempt += 1;
-      if (attempt > maxRetry) {
-        throw e;
-      }
-
-      const status = e.$metadata?.httpStatusCode;
-
-      // As per S3 docs, retry on a 409
-      if (status !== 409) {
-        if (status !== 412) {
-          throw e;
-        }
-        // 412: precondition failed - something is at the destination already.
-
-        if (collOpts.collision === 'unique') {
-          // The request is to move and make the destination file unique.
-          // We do this by appending a ms timestamp to the destination key.
-
-          const ext = `.${destKey.split('.').pop()}`;
-          const destWithoutExt = destKey.slice(0, -ext.length);
-
-          const ts = Date.now().toString(36);
-          destinationKey = `${destWithoutExt}-${ts}${ext}`;
-        } else if (collOpts.collision === 'overwrite') {
-          // eslint-disable-next-line no-await-in-loop
-          const dest = await bucket.head(destKey);
-
-          // version what's there before overwriting it, provide the destination ETag so that we
-          // know we're versioning what we just did a head() of.
-          // eslint-disable-next-line no-await-in-loop
-          const versionResp = await postVersion(context, destKey, 'copy', 'Version created before overwrite', dest.ETag);
-          if (versionResp.status !== 201) {
-            if (versionResp.status !== 412 && versionResp.status !== 409) {
-              throw new StatusCodeError('Failed to version the destination', versionResp.status);
-            }
-          } else {
-            // Creating the version was successful, so we can now copy over the destination.
-
-            const getDestDocId = getDocID(dest);
-
-            // If something is at the destination already, we copy over that file, but keep
-            // the doc ID from the destination as-is so that the destination keeps its history.
-            opts = { ...initialOpts, addMetadata: { 'doc-id': getDestDocId } };
-
-            // Now only copy over the destination if it's still the same as what we did a head() of
-            copyOpts = { IfMatch: dest.ETag };
-          }
-        } else {
-          throw new StatusCodeError('Collision: something is at the destination already', 409);
-        }
-      }
-    }
-  }
-
-  if (move) {
-    const resp = await bucket.remove(srcKey);
-    if (resp.$metadata?.httpStatusCode !== 204) {
-      throw new StatusCodeError(`Failed to remove source: ${srcKey}`, resp.$metadata?.httpStatusCode);
-    }
-  }
-}
-
-async function copyFile(context, srcKey, destKey, move, opts, collOpts) {
-  const copyOpts = { ...opts };
-  if (!move) {
-    copyOpts.addMetadata = { 'doc-id': ulid() };
-  }
-  await copyWithRetry(context, srcKey, destKey, move, copyOpts, collOpts);
-}
-
-/**
- * Copies a document from the source to the destination.
- *
- * @param {import('../support/AdminContext').AdminContext} context context
- * @param {string} src source S3 key
- * @param {import('../support/RequestInfo').RequestInfo} info destination info
- * @param {boolean} move whether to move the source
- * @param {object} opts additional options for the copy operation
- * @param {object} collOpts collision options
- * @returns {Promise<Array<{src: string, dst: string}>>} the copied file details
- */
-export async function copyDocument(context, src, info, move, opts, collOpts) {
-  const dst = getS3KeyFromInfo(info);
-  await copyFile(context, src, dst, move, opts, collOpts);
-  return [{ src, dst }];
-}
-
-/**
- * Copies a folder from the source to the destination.
- *
- * @param {import('../support/AdminContext').AdminContext} context context
- * @param {string} srcKey source S3 key
- * @param {import('../support/RequestInfo').RequestInfo} info destination info
- * @param {boolean} move whether to move the source
- * @param {function(string, string): object} fnOpts additional options for the copy operation.
- * This function is called for each object being copied with source and destination S3 keys as
- * arguments.
- * @param {object} collOpts collision options
- * @returns {Promise<Array<{src: string, dst: string}>>} the copied files
- */
-export async function copyFolder(context, srcKey, info, move, fnOpts, collOpts) {
-  const tasks = [];
-  const destKey = getS3Key(info.org, info.site, info.rawPath);
-
-  if (destKey.startsWith(srcKey)) {
-    throw new StatusCodeError('Destination cannot be a subfolder of source', 400);
-  }
-
-  const bucket = HelixStorage.fromContext(context).sourceBus();
-  (await bucket.list(srcKey)).forEach((obj) => {
-    tasks.push({
-      src: obj.key,
-      dst: `${destKey}${obj.path}`,
-    });
-  });
-
-  if (tasks.length === 0) {
-    // Nothing found at source
-    throw new StatusCodeError('Not found', 404);
-  }
-
-  const copied = [];
-  await processQueue(tasks, async (task) => {
-    const opts = fnOpts(task.src, task.dst);
-    await copyFile(context, task.src, task.dst, move, opts, collOpts);
-    copied.push({ src: task.src, dst: task.dst });
-  });
-  return copied;
-}
 
 /**
  * Copies a resource of a folder to the destination folder. If a folder is
